% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phantom_reductions.R
\name{create_sankey_bump}
\alias{create_sankey_bump}
\title{Create Sankey Bump Function}
\usage{
create_sankey_bump(
  phantom_emissions_flags,
  time,
  ID,
  category,
  value,
  id_instance
)
}
\arguments{
\item{phantom_emissions_flags}{Dataset with dropoff?pickup flags, resulting dataset from phantom_flags function}

\item{time}{Time that each row represents}

\item{ID}{Identifying key}

\item{category}{Categorization of value}

\item{value}{Numeric value of interest}

\item{id_instance}{Specific ID that is visualized}
}
\description{
This function creates sankey bump graphs to visualize phantom reductions
}
\examples{
create_sankey_bump(OP_flags, year, company, funding_category, n, 3)
}
