% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phantom_reductions.R
\name{dropoff_and_pickup}
\alias{dropoff_and_pickup}
\title{Dropoff and Pickup Filter Function}
\usage{
dropoff_and_pickup(phantom_emissions_flags)
}
\arguments{
\item{phantom_emissions_flags}{Dataset with dropoff?pickup flags, resulting dataset from phantom_flags function}
}
\description{
This function filters the dataset to just rows with phantom reductions
}
\examples{
dropoff_and_pickup(OP_flags)
}
\keyword{dropoff}
\keyword{pickup}
