% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phantom_reductions.R
\name{phantom_flags}
\alias{phantom_flags}
\title{Phantom Flags Function}
\usage{
phantom_flags(aggregated_data, time, ID, category, value)
}
\arguments{
\item{aggregated_data}{Dataset}

\item{time}{Time that each row represents}

\item{ID}{Identifying key}

\item{category}{Categorization of value}

\item{value}{Numeric value of interest}
}
\description{
This function flags rows with drop offs or pick ups from the previous time
}
\examples{
OP_flags <- phantom_flags(open_payments, year, company, funding_category, n)
}
\keyword{phantom}
\keyword{reductions}
